частично разнес логику по компонентам
